LayoutTests/platform/gtk/imported/w3c/web-platform-tests/fetch/api/basic/scheme-about.any-expected.txt

CONSOLE MESSAGE: line 10: Fetch API cannot load about:blank. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: Fetch API cannot load about:blank. The URL can’t be shown
CONSOLE MESSAGE: line 10: Fetch API cannot load about:blank. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 27: Fetch API cannot load about:invalid.com. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 27: Fetch API cannot load about:config. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 27: Fetch API cannot load about:unicorn. Cross origin requests are only supported for HTTP.

FAIL Fetching about:blank (GET) is OK promise_test: Unhandled rejection with value: object "TypeError: Cross origin requests are only supported for HTTP."
FAIL Fetching about:blank (PUT) is OK promise_test: Unhandled rejection with value: object "TypeError: Preflight response is not successful"
FAIL Fetching about:blank (POST) is OK promise_test: Unhandled rejection with value: object "TypeError: Cross origin requests are only supported for HTTP."
PASS Fetching about:invalid.com is KO 
PASS Fetching about:config is KO 
PASS Fetching about:unicorn is KO 
